Have receive codepath work
Use go-message to parse internet message emails.

Split code into send and receive functions. Send is used to send
mail through nncp. Receive is used to receive mail, edit the From
header, and then send it to the local MTA.

Use a subfunction to create the receive mode new email address

Add aliases for "send", "receive", and "recv" which actually branch to
execute the correct send or receive function.

log to Stderr since in the receipt case the mail is written to stdout